refactor(desktop-client): Phase 1.4.A — ToolStep 类型从旧 Provider 抽离
变更
----
将 `ToolStep` 接口从 `runtime/TauriRuntimeProvider.tsx` 迁移到
`runtime/contexts/ApprovalProvider.tsx`，与 `PendingApproval` 聚合在
"工具交互" 域下：

* `ApprovalProvider.tsx`：新增 `export interface ToolStep`
* `TauriRuntimeProvider.tsx`：删除原定义，改为
  `import type + export type` 的 re-export 形式，保持旧外部
  `import { type ToolStep } from '../TauriRuntimeProvider'` 继续可用
* `components/assistant-ui/thread.tsx`：
  `import { type ToolStep }` 指向 `ApprovalProvider`
* `components/assistant-ui/context-tool-group.tsx`：同步修改

背景
----
Phase 1.4 要彻底下线旧 `TauriRuntimeProvider`；先把 `ToolStep` 搬走
可在不引入功能变更的前提下，为后续删旧文件解除类型依赖。本次不涉及
`ChatTabTauri` / `useExperimentalRuntime` 的删除 —— 这些改动会造成
`outboundCommand` / `engineReadyKey` 能力暂时丢失（尚未在新 runtime
迁移），需要先与上游决策后再推进。

测试
----
全量 vitest：858 tests / 828 passed，与 `97d1d6d7` 基线完全一致，
0 回归。
